// src/app/model/fluidLevelMeasurementResponse.ts
/**
 * Oil and Gas Reporting System (OGRS-API)
 * This is an api  to allows users to obtain infomation about the oil and gas wells database and generating a punch of related reposts such as wells count, insurance as well as production allocation funcationality
 *
 * OpenAPI spec version: 3.0
 * Contact: mohamed.wafa770@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Model containing FluidLevelMeasurement object info for returns (get requests)
 */
export interface FluidLevelMeasurementResponse { 
    id?: number;
    date?: Date;
    intervals?: string;
    flType?: FluidLevelMeasurementResponse.FlTypeEnum;
    fluidLevel?: number;
    pumpDepth?: number;
    liqPercentage?: number;
    pumpFillage?: number;
    pumpSubmerge?: number;
    card?: FluidLevelMeasurementResponse.CardEnum;
    remarks?: string;
}
export namespace FluidLevelMeasurementResponse {
    export type FlTypeEnum = 'Static' | 'Dynamic';
    export const FlTypeEnum = {
        Static: 'Static' as FlTypeEnum,
        Dynamic: 'Dynamic' as FlTypeEnum
    };
    export type CardEnum = 'a' | 'b' | 'c';
    export const CardEnum = {
        A: 'a' as CardEnum,
        B: 'b' as CardEnum,
        C: 'c' as CardEnum
    };
}